Lagra konversationer på fil
Lagra konversationer på fil fungerar nu, imorgon skall vi implementera historik samt buzz

+++ TDDD49/Model/FileWriter.cs
@@ -20,7 +20,9 @@ namespace TDDD49.Model
             Directory.CreateDirectory(@"c:\TDDD49STORAGE");
 
             //Skapar fil
-            File.WriteAllText(@"c:\TDDD49STORAGE\conversations.json", String.Empty);
+            if (!File.Exists(@"c:\TDDD49STORAGE\conversations.json")){
+                File.WriteAllText(@"c:\TDDD49STORAGE\conversations.json", String.Empty);
+            }
 
             if (File.ReadAllText(@"c:\TDDD49STORAGE\conversations.json") == String.Empty)
                 {
@@ -37,26 +39,9 @@ namespace TDDD49.Model
 
         public void WriteToFile(JObject jsonObj) 
         {
-            //ska ta in jsonobj
-            //ska hämta hem nuvarande konversation från lokal fil
-            //ska parta konversation till jsonobj
-            //ska lägga till jsonobj till hämtad json
-            //ska skriva json till lokal fil
-
-            //List<List<Message>> Data = JsonConvert.DeserializeObject<List<List<Message>>>(conversations.First.ToString());
-            //JObject oldConversations = JObject.Parse(File.ReadAllText(@"c:\TDDD49STORAGE\conversations.json"));
 
             JArray arrayOfConvos = (JArray)conversations["conversations"];
-            JArray aConvo;
-            if ((JArray)arrayOfConvos.Last == null)
-            {
-                aConvo = new JArray();
-                arrayOfConvos.Add(aConvo);
-              }
-            else
-            {
-                aConvo = (JArray)arrayOfConvos.Last;
-            }
+            JArray aConvo = (JArray)arrayOfConvos.Last;        
             
             aConvo.Add(jsonObj);
 
@@ -64,17 +49,13 @@ namespace TDDD49.Model
 
             File.WriteAllText(@"c:\TDDD49STORAGE\conversations.json", conversations.ToString());
 
-
         }
-        /**
+        
         public void InitConversation()
         {
-            //funktion ska skapa ny konversation (vid behov)
-            if (!conversations.Last.HasValues)
-            {
-                conversations.AddAfterSelf(new JProperty("conversation", new JArray()));
-            }
-        }*/
-        
+            JArray arrayOfConvos = (JArray)conversations["conversations"];
+            arrayOfConvos.Add(new JArray());
+        }
+       
     }
 }
